// database/prisma-thinkthank/generated/zod/inputTypeSchemas/RegisteredDomainUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';

export const RegisteredDomainUncheckedCreateInputSchema: z.ZodType<Prisma.RegisteredDomainUncheckedCreateInput> = z.object({
  id: z.string().cuid().optional(),
  slug: z.string(),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional(),
  expiresAt: z.coerce.date(),
  projectId: z.string(),
  domainId: z.string().optional().nullable()
}).strict();

export default RegisteredDomainUncheckedCreateInputSchema;